Use an incrementing serial to name DBPortPool MBean
Object.hashCode is not guaranteed to be unique even for
objects that are live at the same time.
This may be because the garbage collector can move objects
around and the hashCode is based on the object's location
the first time hashCode is called, not its current location.

This fixes a failure to register the mbean due to name
collisions caused by hashCode collisions.

The incrementing serial is probably more friendly than
a random hashCode number anyhow.