A refactoring of the driver's core message handling capabilities is necessary to prepare for the new wire protocol message types (OP_COMPRESSED and OP_MSG) in MongoDB 3.6.
In short, higher-level message handling needs to be centralized within InternalConnection so that the new
message types can be handled in a single place rather than spread out in the various Protocol
implementations.

Details:

* CommandMessage becomes an abstract base class for SimpleCommandMessage and BaseWriteCommandMessage
* RequestMessage adds the capability to remember the EncodingMetadata rather than just return it from encode.
  This will allow WriteCommandProtocol to handle batch splitting even though it no longer is responsible
  directly for encoding the message.
* InternalConnection gets sync and and async sendAndReceive methods which take a CommandMessage and
  handle both sending the command message and receiving the response.
* InternalStreamConnection implements the methods and takes over responsibility for sending command events
  when those methods are used.
* CommandProtocol and WriteCommandProtocol are refactored to use these new methods, and so no longer are
  responsible for encoding the command or sending command events

One functional change is made to InternalStreamConnection: it no longer is capable of concurrent use, since
that would have complicated the implementation of the sendAndReceive methods.  The justification is that
this capability was not actually used by the driver, as its only purpose is to enable message pipelining and
the driver ended up not supporting that use case.  It can always be added back if necessary.

After the refactoring the InternalConnection interface is a bit of a hybrid.  It has the new sendAndReceive
methods that take a CommandMessage.  And it still has the sendMessage  methods that take a byte buffer and
are used by the other Protocol implementation classes. When server versions that rely on the old wire protocol
are no longer supported then these class and methods can just be removed and the driver will be much simpler.
But for now there's not a good reason to spend time refactoring any further.