TypeData no longer validates data.
With the introduction of PropertyCodecProviders the
TypeData class no longer should validate the data it sees
for Collections or Maps.

This changes behavior for "invalid Maps" to a possible
runtime CodecConfigurationException. However, it
allows users to register their own PropertyCodecProviders
to handle these types if desired.

The side effect of this change is @BsonIgnore and
conventions can now be used to ignore properties
regardless of validity.

JAVA-2620