JAVA-2757: Drop index by keys when keys are provided
When the application drops an index using a method that takes the index
keys, rather than the name, the driver drops the index by those keys
rather than constructing a name from the keys and dropping it by that
name.

This allows the dropping of index by keys even when the index was
created with a name other than the default name.