JAVA-2859: Remove extra server session pool
Now that the extra server session pool is gone, it exposed an error
in a ClientSession test which creates two client sessions in a Spock
where clause.  Since both sessions are created by Spock in advance of
the test executions, it means that the cleanup check for in-use sessions
triggers after the first test execution completes and asserts that
there is one in-use session.