Add support for 'dropDatabase' and 'rename' operation types
This commit adjusts the ChangeStreamDocument so that it takes in
a BsonDocument for the namespace instead of a MongoNamespace object.
This is to combat the issue of the dropDatabase operation responding
with a ChangeStreamDocument that only contains the databaseName and
not the collectionName. Tests were updated to account for this as
well. Additionally, a test for the rename operation type was added.

JAVA-2966