Deprecate MongoIterable#forEach(Block<? super TResult> block)
With the addition of the default method Iterable#forEach(Consumer)} in
Java 8, MongoIterable#forEach(Block<? super TResult> block) results in
an error at the point of call of forEach when using a lambda, which can
only be resolved by casting to either Block or Consumer. Rather than
keep that unfriendly API, we deprecate MongoIterable#forEach so that
it can be removed in the next major release of the driver (which will
required Java 8).

JAVA-3046