Use a failpoint to force write concern errors in drop database tests
Starting in MongoDB 4.0, the technique of using a write concern with a
"w" value that is greater than the number of replica set members in
order to force a write concern fails for the drop database command.
Also starting in MongoDB 4.0, there is a failpoint available that allows
tests to force a write concern error.  This patch updates the drop
database write concern error tests to use the failpoint when the server
version is >= 4.0

JAVA-2868