JAVA-3328: remove the isEmpty check when making bulk write updates
Handling empty object detection is proving to have too much code overhead.
The previous solution ignored pipe and so is incomplete. Supporting pipe
requires similar delegating wrappers for BsonReader.

Since the server checks for this condition anyway, it's not a big loss
to lose the validation point on this client side, given its significant
performance downside.