Made CodecConfigurationException error message more explicit
It just took me several hours to figure out the fact that the problem was not that no public constructor existed (because it did), but that the problem was that the BSON deserializer requires the class to have an *empty* constructor with *no arguments*. Once I figured that out, I was able to solve the problem within less than a minute. I hope that this more explicit error message helps other people who run into the same issue with their first time trying to deserialize BSON objects.