Sync to latest change stream tests
There is one behavioral change in these tests: to treat CursorNotFound errors as resumable.
The reason there is no corresponding change in the driver is because it was already doing this,
incorrectly it turns out.  But now it turns it that it is the correct thing to do.

JAVA-3781
JAVA-3776