A few improvements to the code
* ReferenceCounted.getCount documentation was changed.
It is documented to start at 1, which is with accordance to
what the current driver code expects. The common usage pattern is as follows:
ReferenceCounted resource = new ...;
//note that resource.retain() is not called because we expect the resource is created with counter 1
try {
  //use resource witho;
} finally {
  resource.release();
}
Changing implementations of ReferenceCounted.getCount to start with 0 would result in
resource.release() throwing IllegalStateException("Attempted to decrement the reference count below 0").
As long as the method ReferenceCounted.getCount is exposed, its behavior
is a part of the ReferenceCounted contract.

* Some fields in AsyncQueryBatchCursor were made final.

* Reduced the amount of garbage produced by QueryResult.getCursor.