Disable test for change stream projecting out the _id field
There's an integration test that asserts an exception is thrown when a change stream
projects out the _id field.  Running against 3.6 and 4.0, this condition is checked
in the driver itself while processing the results.  But in 4.4 the server started
replying with an error on the getMore in this scenario.  A newer driver would know not to
retry in this case, because newer drivers implemented an allow list for resumeable change
stream errors when the server version. But older drivers don't have that logic, so
it will treat this command error as retryable and never actually report the error.

JAVA-3942