Fix connection reference counting in AsyncQueryBatchCursor
There are two things that need to be addressed about this commit:
"structured" reference counting, self-managing ReferenceCounted objects.

"Structured" reference counting.
The idea is to try and get our reference counting code closer to
following the structured programming approach
(https://en.wikipedia.org/wiki/Structured_programming)
while fixing a bug: AsyncQueryBatchCursor does not always release a connection
when it is not needed anymore.

In simple words, the following
is desirable: given a place in code where retain/release is called,
it should be simple to tell where the corresponding release/retain is called.

As a side note, structured programming applied to concurrent code
is called structured concurrency. This idea is realized in Kotlin:
https://kotlinlang.org/docs/reference/coroutines/basics.html#structured-concurrency,
https://elizarov.medium.com/structured-concurrency-722d765aa952
(the article is written by the author of Kotlin coroutines functionality).

Self-managing ReferenceCounted objects.
The idea of reference counting breaks when a resource may be used without
the need to have a reference to it. For example, consider the following code

asyncConnection.commandAsync((result, exception) ->
    //process result without touching the connection
)

On one hand, the callback does not store a reference to asyncConnection,
so the caller of the commandAsync method does not see a reason to retain asyncConnection;
on the other hand, the asyncConnection should not be released until
the asynchronous logic completes, otherwise we may never receive a result.
Clearly, asyncConnection must be retained by someone, but not by the caller who
cannot deduce this based only on thinking about references
(and this is all he should be thinking about to preserve mental health).

To cope with the situation, we can put the burden of retaining and releasing
asyncConnection on the connection itself. And this is the approach
implemented in this commit.

JAVA-3907