Start releasing connectionSource in AsyncQueryBatchCursor.close
A reader may not like either of / both the following methods:
connectionSource, releaseConnectionSourceIdempotently.
So it is worth explaining why they were introduced.

connectionSource
The original code stored a reference to connectionSource in an instance field
without retaining the object in some cases (specifically, if cursor is null).
The code in then tried to avoid using such an unretained and, therefore,
potentially unusable object.
The new code does not make things better in this respect, but it makes
it impossible for a writer to try and use a connectionSource
that has not been retained in the constructor. The new code also
checks the state and throws an exceptions more clearly stating
that the problem is in the code of the AsyncQueryBatchCursor.

releaseConnectionSourceIdempotently
Code releasing connectionSource is sprinkled all over AsyncQueryBatchCursor.
All my attempts to have the connectionSource released only by the close method
did not succeed - there were always numerous tests that start failing as a result.
This class is in need to be refactored such that all resources
(connectionSource, server cursor, etc.) are released only as a result of calling
the close method. This method may then be called either
from the AsyncQueryBatchCursor class itself, or externally.
Such refactoring is not trivial, and will also affect some tests,
that depend on the internals of the AsyncQueryBatchCursor
(definitely driver-core/src/test/unit/com/mongodb/internal/operation/AsyncQueryBatchCursorSpecification.groovy
and some of its neighbours)
In other words, the method releaseConnectionSourceIdempotently works around
problems arising from the structure of the current code,
and allows makes it possible to release connectionSource by calling the close method.

JAVA-3907