Release connections early in AsyncQueryBatchCursor
Just like in the ResourcePreservingCallback.onResult,
callbacks in AsyncQueryBatchCursor must release connection
before doing the callback logic, provided that the connection
is not used by that logic.

It also worth pointing out that the method AsyncConnectionSource.getConnection
does not decorate callbacks to release connections after running the callback.
This behavior interferes with the idea of structured reference counting.
It forces the user to think about the fact, that AsyncConnectionSource.getConnection
is the "starting" point of a lifetime of a reference counted connection and
has to be treated differently by releasing the connection despite the user
not retaining it.

AsyncConnectionSource.getConnection could behave as follows:

getConnection((connection, t) -> {
    //use connection, no need to release unless you explicitly retain
}//getConnection releases the connection automatically after return from the callback
)

acting effectively as doWithConnection(Consumer<Connection> action),
but allowing a connection to survive returning from the action.consume method
if the method stored a reference to the connection somewhere and retained it.
It's not possible to change the current behavior without changing all places
where the method is used, which is a no-go.
Alternatively, a new method
AsyncConnectionSource.doWithConnection(Consumer<Connection> action)
may be introduced.

JAVA-3907