Introduce self-managing reference counting to AsyncConnection and AsyncConnectionSource
The idea of reference counting breaks when a resource may be used without
the need to have a reference to it. For example, consider the following code

asyncConnection.commandAsync((result, exception) ->
    //process result without touching the connection
)

On one hand, the callback does not store a reference to asyncConnection,
so the caller of the commandAsync method does not see a reason to retain asyncConnection;
on the other hand, the asyncConnection should not be released until
the asynchronous logic completes, otherwise we may never receive a result.
Clearly, asyncConnection must be retained by someone, but not by the caller who
cannot deduce this based only on thinking about references
(and this is all he should be thinking about to preserve mental health).

To cope with the situation, we can put the burden of retaining and releasing
asyncConnection on the connection itself. And this is the approach
implemented in this commit.

JAVA-3978