Update UnifiedCrudHelper to involve ChangeStreamDocumentCodec in the test process
This approach deserializes a ChangeStreamDocument from the incoming data
and then serializes it in a BsonDocument. This approach works so long as
the server responds with [] (an empty array) instead of null when specifying
an empty UpdateDescription.truncatedArrays field.

If the server starts returning null instead of [],
we will need to modify ChangeStreamDocumentCodec so that it encodes
empty List values as BSON null. This is because UpdateDescription
represents null values as an empty List
in the UpdateDescription.getTruncatedArrays method.

If the server starts using both null and []
in different responses to report an empty UpdateDescription.truncatedArrays field,
then there no change in the ChangeStreamDocumentCodec may help us
fixing the problem, and we will have to either undo this commit,
or change the way UnifiedTest compares expectResult with the document
sent by the server.

JAVA-3990