Ignore CancelledKeyException in AsynchronousTlsChannelGroup#loop
In testing we've observed that SelectionKey#interestOps can sometimes
throw a CancelledKeyException when channels are being closed forcefully
by the server in response to operations with failpoints set on them.

CancelledKeyException is already ingored elsewhere in
AsynchronousTlsChannelGroup, and it seems it has to also be ignore
in the processPendingInterests method.

JAVA-4024