Refactor `DefaultConnectionPool` to use single worker thread when checking out asynchronously
When implementing https://github.com/mongodb/mongo-java-driver/commit/73e30857eed2f1d02ff7e9a6e5028d557a1cf4cf
I correctly explained the deadlock situation:
1) An async task that created a non-open connection and tries to open it (let us call it an "open task")
may block another async task that tries to create a connection (let us call it a "get task")
if `maxSize` is reached.
2) If both tasks are submitted to the same queue such that the get task is ahead of the open task, then
such a get task naturally blocks the open task while being blocked by the open task.

The solution to the problem I used is to submit tasks to different queues, which,
because both tasks are blocking, required introducing one more async worker thread.

However, there is another approach to solving the deadlock problem,
which did not occur to me at that time: prevent the get task to be ahead of the open task
in the queue. This may happen due to the optimization in the `getAsync` method, which
tries to check out synchronously if an open connection is available,
i.e., if checking it out requires no blocking. It is a nice optimization that I
tried to preserve (it was there before my changes) when implemented JAVA-3927.
However, this optimization opens up opportunities for a get task being blocked
by an open task to be ordered ahead of the open task in the queue.

If I realized that previously, I would have sacrificed this optimization
to eliminate the deadlock instead of introducing another queue and a thread.
Fortunately, the it occurred to me now.

JAVA-4146